+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Handle URL fragments in redirects (#9).
 
+* `url_check()` now also handles package tarballs. This is useful to check
+  the URL in the package vignettes.
+
 # urlchecker 1.0.1
 
 * Gábor Csárdi is now the maintainer.